[arccore,collections] Met dans des fichiers différents les tests 'Array2' et 'Array'.
