+++ arccore/src/collections/tests/CMakeLists.txt
@@ -1,6 +1,7 @@
 arccore_add_component_test_executable(collections
   FILES
   TestArray.cc
+  TestArray2.cc
   )
 
 target_link_libraries(arccore_collections.tests PUBLIC arccore_collections GTest::GTest GTest::Main)
